+++ CHANGELOG.md
@@ -15,6 +15,7 @@
         * `RSAPadding`
         * `SymmetricEncryptionAlgorithm` and its nested implementations
 * Fix project setup on non-macOS build hosts
+* Limit Keystore operations with `limitedParallelism` from `kotlinx.coroutines`
 
 ### 3.18.2 / Supreme 0.10.2
 * Fixes